+++ src/structures/src/String/strmod.c
@@ -20,7 +20,7 @@ void strmod( SAC_ND_PARAM_out( new_nt, string),
   SAC_ND_DECL__DATA( new_nt, string, )
   SAC_ND_DECL__DESC( new_nt, )
 
-  RANGECHECK( pos, 0, strlen( SAC_ND_A_FIELD( old_nt)) - 1,
+  RANGECHECK( (size_t)pos, 0, strlen( SAC_ND_A_FIELD( old_nt)) - 1,
               SAC_ND_A_FIELD( old_nt));
 
   if (SAC_ND_A_RC( old_nt) == 1) {
@@ -47,7 +47,7 @@ void strmod( SAC_ND_PARAM_out_rc( string, new),
 {
   SAC_ND_DECL_RC(string, new);
 
-  RANGECHECK(pos, 0, strlen(old)-1, old);
+  RANGECHECK((size_t)pos, 0, strlen(old)-1, old);
   
   if (SAC_ND_A_RC(old)==1) {
     new=old;
